BREAKING: rearrange provider s modules without claddification dirs
