+++ lib/spaces/providers/docker/divisions/images.rb
@@ -1,0 +1,11 @@
+module Providers
+  module Docker
+    class Images < ::ProviderAspects::Images
+
+      def packing_artifact
+        super.join("\n")
+      end
+
+    end
+  end
+end
