+++ lib/spaces/providers/docker/patch/image.rb
