+++ lib/spaces/providers/packer/divisions/configuration.rb
@@ -1,0 +1,9 @@
+module Providers
+  module Packer
+    class Configuration < ::ProviderAspects::Configuration
+
+      def packing_artifact ;end
+
+    end
+  end
+end
