+++ lib/spaces/providers/packer/divisions/execution.rb
@@ -1,0 +1,9 @@
+module Providers
+  module Packer
+    class Execution < ::ProviderAspects::Execution
+
+      def packing_artifact ;end
+
+    end
+  end
+end
