+++ lib/spaces/providers/packer/divisions/modules.rb
@@ -1,0 +1,14 @@
+module Providers
+  module Packer
+    class Modules < ::ProviderAspects::Modules
+
+      def packing_artifact_for(language)
+        {
+          type: 'shell',
+          inline: division.send(language).inline
+        }
+      end
+
+    end
+  end
+end
