+++ lib/spaces/providers/packer/divisions/permission.rb
@@ -1,0 +1,19 @@
+module Providers
+  module Packer
+    class Permission < ::ProviderAspects::Permission
+
+      delegate [:recursion, :ownership, :file, :mode] => :division
+
+      def packing_artifact
+        {
+          type: 'shell',
+          inline: [
+            ("chown #{recursion} #{ownership} #{file}" if ownership),
+            ("chmod #{recursion} #{mode} #{file}" if mode)
+          ].compact
+        }
+      end
+
+    end
+  end
+end
