--- lib/spaces/providers/packing/packer/configuration.rb
@@ -1,9 +1,0 @@
-module Providers
-  class Packer < ::ProviderAspects::Provider
-    class Configuration < ::ProviderAspects::Configuration
-
-      def packing_artifact ;end
-
-    end
-  end
-end
