--- lib/spaces/providers/packing/packer/execution.rb
@@ -1,9 +1,0 @@
-module Providers
-  class Packer < ::ProviderAspects::Provider
-    class Execution < ::ProviderAspects::Execution
-
-      def packing_artifact ;end
-
-    end
-  end
-end
