--- lib/spaces/providers/packing/packer/lxd/images.rb
@@ -1,11 +1,0 @@
-module Providers
-  class Packer < ::ProviderAspects::Provider
-    module Lxd
-      class Images < ::ProviderAspects::Images
-
-        def packing_artifact; division.to_h ;end
-
-      end
-    end
-  end
-end
