--- lib/spaces/providers/provisioning/terraform/docker/service_tasks.rb
@@ -1,17 +1,0 @@
-module Providers
-  class Docker < ::ProviderAspects::Provider
-    class ServiceTasks < ::ProviderAspects::ServiceTasks
-
-      def connection_stanza_for(binding)
-        connect&.map do |c|
-          %(
-            provisioner "local-exec" {
-              command = "docker exec #{binding.environment_variables}  #{blueprint_identifier} #{c} "
-            }
-          )
-        end.join
-      end
-
-    end
-  end
-end
