--- lib/spaces/providers/provisioning/terraform/lxd/volume.rb
@@ -1,34 +1,0 @@
-module Providers
-  class Lxd < ::ProviderAspects::Provider
-    class Volume < ::ProviderAspects::Volume
-
-      def device_stanzas
-        %Q(
-          device {
-            name = "#{volume_name}"
-            type = "disk"
-            properties = {
-              path = "#{destination}"
-              source = "#{volume_name}"
-              pool = "#{pool_name}"
-            }
-          }
-        )
-      end
-
-      def resolution_stanzas_for(_)
-        %Q(
-          resource "lxd_volume" "#{volume_name}" {
-            name = "#{volume_name}"
-            pool = "#{pool_name}"
-            remote = "lxd-server"
-          }
-        )
-      end
-
-      def volume_name; "#{blueprint_identifier}-#{source}" ;end
-      def pool_name; "#{source}-pool" ;end
-
-    end
-  end
-end
